src/containers/videoclip-embedding/requirements.txt
torch>=2.1.0
transformers>=4.35.0
boto3>=1.28.0
opensearch-py>=2.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.0
pydantic>=2.4.0
python-jose>=3.3.0
tenacity>=8.2.0
requests-aws4auth>=1.2.0
Pillow>=10.0.0